feat: 채팅방 나간 후 멤버목록 새로 불러오기
