+++ src/api/channel.ts
@@ -6,6 +6,7 @@ import {
 } from '../@types/channel';
 import { checkChannelName } from '../utils';
 import instance from './axios';
+import { getUser } from './user';
 
 export interface CreateChannelBody {
   name: string;
@@ -70,4 +71,20 @@ export const participateChannel = async (chatId: string) => {
     newChatId,
   );
   return response.data;
+};
+
+export const getMemberData = async (users: string[]) => {
+  const userListData = [];
+  for (const id of users) {
+    const response = await getUser(id);
+    if (response) {
+      const { name, picture } = response;
+      userListData.push({
+        id,
+        name,
+        picture,
+      });
+    }
+  }
+  return userListData;
 };
